[chore-#38] 패키지명 이니셜 소문자로 변경
