--- src/main/java/com/bss/bssserverapi/domain/Auth/dto/LoginUserResWithCookieDto.java
@@ -1,22 +1,0 @@
-package com.bss.bssserverapi.domain.Auth.dto;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.http.ResponseCookie;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
-public class LoginUserResWithCookieDto {
-
-    private ResponseCookie cookie;
-    private LoginUserResDto loginUserResDto;
-
-    @Builder
-    public LoginUserResWithCookieDto(final ResponseCookie cookie, final LoginUserResDto loginUserResDto) {
-
-        this.cookie = cookie;
-        this.loginUserResDto = loginUserResDto;
-    }
-}
