--- src/main/java/com/bss/bssserverapi/domain/Auth/repository/AuthJpaRepository.java
@@ -1,11 +1,0 @@
-package com.bss.bssserverapi.domain.Auth.repository;
-
-import com.bss.bssserverapi.domain.Auth.Auth;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface AuthJpaRepository extends JpaRepository<Auth, String> {
-
-    Optional<String> findRefreshTokenByUserId(final String userId);
-}
