--- src/main/java/com/bss/bssserverapi/domain/User/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.bss.bssserverapi.domain.User.repository;
-
-import com.bss.bssserverapi.domain.User.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByUserId(final String userId);
-    boolean existsByUserId(final String userId);
-}
