+++ src/main/java/com/bss/bssserverapi/domain/auth/filter/ExceptionHandlerFilter.java
@@ -1,0 +1,30 @@
+package com.bss.bssserverapi.domain.auth.filter;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.web.filter.OncePerRequestFilter;
+import org.springframework.web.servlet.HandlerExceptionResolver;
+
+import java.io.IOException;
+
+public class ExceptionHandlerFilter extends OncePerRequestFilter {
+
+    private final HandlerExceptionResolver handlerExceptionResolver;
+
+    public ExceptionHandlerFilter(HandlerExceptionResolver handlerExceptionResolver) {
+
+        this.handlerExceptionResolver = handlerExceptionResolver;
+    }
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+
+        try {
+            filterChain.doFilter(request, response);
+        } catch (Exception e){
+            handlerExceptionResolver.resolveException(request, response, null, e);
+        }
+    }
+}
