+++ src/main/java/com/bss/bssserverapi/domain/auth/filter/GlobalAccessDeniedHandler.java
@@ -1,0 +1,35 @@
+package com.bss.bssserverapi.domain.auth.filter;
+
+import com.bss.bssserverapi.global.exception.ErrorCode;
+import com.bss.bssserverapi.global.exception.GlobalException;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.web.servlet.HandlerExceptionResolver;
+
+import java.io.IOException;
+
+public class GlobalAccessDeniedHandler implements AccessDeniedHandler {
+
+    private final HandlerExceptionResolver resolver;
+
+    public GlobalAccessDeniedHandler(HandlerExceptionResolver resolver) {
+
+        this.resolver = resolver;
+    }
+
+    @Override
+    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException, IOException {
+
+        resolver.resolveException(
+                request,
+                response,
+                null,
+                new GlobalException(HttpStatus.FORBIDDEN, ErrorCode.UNAUTHORIZED)
+        );
+    }
+
+}
