+++ src/main/java/com/bss/bssserverapi/domain/auth/repository/AuthJpaRepository.java
@@ -1,0 +1,11 @@
+package com.bss.bssserverapi.domain.auth.repository;
+
+import com.bss.bssserverapi.domain.auth.Auth;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface AuthJpaRepository extends JpaRepository<Auth, String> {
+
+    Optional<String> findRefreshTokenByUserId(final String userId);
+}
