+++ src/main/java/com/bss/bssserverapi/domain/auth/repository/AuthRepository.java
@@ -1,0 +1,13 @@
+package com.bss.bssserverapi.domain.auth.repository;
+
+import java.util.Date;
+import java.util.Optional;
+
+public interface AuthRepository {
+
+    void save(final String userId, final String token, final Date expiredDate);
+
+    Optional<String> refreshToken(final String userId);
+
+    void delete(final String userId);
+}
