+++ src/main/java/com/bss/bssserverapi/domain/auth/repository/AuthRepositoryImpl.java
@@ -1,0 +1,37 @@
+package com.bss.bssserverapi.domain.auth.repository;
+
+import com.bss.bssserverapi.domain.auth.Auth;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import java.util.Date;
+import java.util.Optional;
+
+@Repository
+@RequiredArgsConstructor
+public class AuthRepositoryImpl implements AuthRepository{
+
+    private final AuthJpaRepository authJpaRepository;
+
+    @Override
+    public void save(final String userId, final String refreshToken, final Date expiredDate) {
+
+        authJpaRepository.save(Auth.builder()
+                .userId(userId)
+                .refreshToken(refreshToken)
+                .expiredTime(expiredDate)
+                .build());
+    }
+
+    @Override
+    public Optional<String> refreshToken(final String userId) {
+
+        return authJpaRepository.findRefreshTokenByUserId(userId);
+    }
+
+    @Override
+    public void delete(final String userId) {
+
+        authJpaRepository.deleteById(userId);
+    }
+}
