+++ src/main/java/com/bss/bssserverapi/domain/user/User.java
@@ -1,0 +1,32 @@
+package com.bss.bssserverapi.domain.user;
+
+import com.bss.bssserverapi.global.common.DateTimeField;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name = "users")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User extends DateTimeField {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true, nullable = false, length = 255)
+    private String userId;
+
+    @Column(nullable = false, length = 255)
+    private String password;
+
+    @Builder
+    public User(final String userId, final String password) {
+
+        this.userId = userId;
+        this.password = password;
+    }
+}
