+++ src/main/java/com/bss/bssserverapi/domain/user/dto/SignupUserResDto.java
@@ -1,0 +1,19 @@
+package com.bss.bssserverapi.domain.user.dto;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PACKAGE)
+public class SignupUserResDto {
+
+    private String userId;
+
+    @Builder
+    public SignupUserResDto(final String userId) {
+
+        this.userId = userId;
+    }
+}
